feat: read in the exe arguments from SPEChpc
SPEChpc generates a run directory with a `control` file that has the
arguments for the benchmark application. This is now copied back to the
stage directory so that the arguments may be read into the executable
options before ReFrame runs the benchmark.

If the executable options are already defined, this step is ignored, so
that smaller, custom benchmarks can still easily be defined.