feat: poll perf counters periodically
One of the repeated bugs encountered on the cluster is that the magic
telephone number 4-294-967-295 kept being recorded as the energy
measurement. Since perf is using RAPL under the hood, and since RAPL
uses micro-joules as the energy estimate, it is very likely that during
the benchmark this counter ticks over the u32 maximum, resulting in the
magic telephone number.

The fix this patch introduces is to probe periodically, since perf seems
to reset the counter after each measurement. This way, we can keep
(micro) Joule estimates sane.