feat: database query fetcher
Records the rough start and endtime of the job and makes a query to the
database to fetch the time series. Currently it still only returns 0
from the reframe performance function.

Added a debug mode that can be enabled with the
`SRFM_PROMETHEUS_DEBUG_ONLY` environment variable being set to any
non-zero value.