feat: add time series capture
By making the time series dictionary a variable of the test it gets
saved to the report JSON. We therefore get all of the benchmark and
energy data in one place to be analyzed later.

To make the database power series 'capturable' requires it to be called
with the current `self` state of the regression test (closure capture
did not seem to work).

For the regression value, the power from the Board Management Controller
(BMC) is integrated with the trapezoid rule to recover (an estimate of)
the energy usage.