fix: allow job parallelization
Since we pass a `--nodelist`, unless `--nodes` is also given to the
scheduler, reframe will always allocate all nodes to each job even if we
only want to run it on one. By explicitly adding the `--nodes` option to
the number of nodes in the regression test, we get parallelization for
free.