feat: specify DAG for builds
We need to build all the benchmarks serially. To achieve this within
reframe, we can do this by making each build step a fixture of the
next benchmark. This means we still get some concurrency, and some
benchmarks will start to run whilst the others are compiling, but all
concurrency problems in compilation should be elided.