feat: multi-node perf aggregation
Some outstanding things that still need doing include correctly parsing
the output from the perf commands on the different nodes.

This commit attempts to solve the nigh-impossible by dissecting and
sewing the launch command back together. It does this in order to ensure
a single instance of `perf stat` is running on each node, irrespective
of the order that MPI assigns tasks. To achieve this, it modified the
launcher script that ReFrame generates to launch

        mpirun \
                --host $HOST1 -n 1 perf stat ... ./a.out : \
                --host $HOST2 -n 1 perf stat ... ./a.out : \
                -n $REMAINING ./a.out

This relies on the "unusual job" launch syntax of MPI and is (from what
I can tell) independent of the MPI implementation. The `$REMAINING`
variable is the total number of tasks minus the number of tasks that
were set running under `perf stat`, so that the occupation is the same.

This might introduce some topology problems in the NUMA domains, but
life is too short to diagnose and fix that.